+++ src/main/java/com/main/easyFix/customer/CustomerRepository.java
@@ -9,8 +9,8 @@ import java.util.Optional;
 
 @Repository
 @Transactional(readOnly = true)
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-  Optional<Customer> findById(Long id);
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+  Optional<Customer> findById(int id);
   Optional<Customer> findByEmail(String email);
   List<Customer> findAll();
 }
